DATAJPA-1535 - When deleting an entity we no longer merge it if it doesn't exist in the database.
The merge happened to allow deleting of unmanaged entities.
As a side effect it triggered an insert for new (and thereby also unmanaged) entities.
By checking if the entity exists we avoid this now.

The check should NOT trigger an additional select since the select would have happened in the past anyway.

Original pull request: #378.

+++ src/main/java/org/springframework/data/jpa/repository/support/SimpleJpaRepository.java
@@ -168,6 +168,19 @@ public class SimpleJpaRepository<T, ID> implements JpaRepositoryImplementation<T
 	public void delete(T entity) {
 
 		Assert.notNull(entity, "The entity must not be null!");
+
+		Object id = entityInformation.getId(entity);
+		if (id == null) {
+			return;
+		}
+
+		T existing = em.find(entityInformation.getJavaType(), id);
+
+		// if the entity to be deleted doesn't exist, delete is a NOOP
+		if (existing == null) {
+			return;
+		}
+
 		em.remove(em.contains(entity) ? entity : em.merge(entity));
 	}
